backend/utils/middleware/cors_middleware.py
# your_project_name/middleware.py

from starlette.middleware.base import BaseHTTPMiddleware
from starlette.responses import JSONResponse

class CORSMiddleware(BaseHTTPMiddleware):
    async def dispatch(self, request, call_next):
        response = await call_next(request)
        response.headers['Access-Control-Allow-Origin'] = '*'
        response.headers['Access-Control-Allow-Methods'] = 'GET, POST, OPTIONS'
        response.headers['Access-Control-Allow-Headers'] = 'Content-Type'
        return response
